requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
anyio==4.4.0
    # via httpx
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via python-telegram-bot
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
logtail-python==0.3.0
    # via -r requirements.in
msgpack==1.0.8
    # via logtail-python
python-telegram-bot==21.4
    # via -r requirements.in
requests==2.32.3
    # via logtail-python
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
urllib3==2.2.2
    # via requests
